+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#184](https://github.com/green-code-initiative/creedengo-java/issues/184) GCI99 - Add new Java rule, avoid CSV format, prefer Apache Parquet
+
 ### Changed
 
 - [#69](https://github.com/green-code-initiative/creedengo-java/issues/69) correction of NullPointer in GCI79 rule + technical refactoring of GCI79